+++ docs/release_notes.rst
@@ -9,6 +9,11 @@ v0.3.15
 * Support for TrueBeam v3.0 MPC has been added. Thanks to `Ash Cullen
   <https://github.com/ash-cullen/>`_ for this contribution.
 
+* For file upload pumps, there is a new "Use File Modified Time" option.
+  Setting this to True will set the ``work_started`` date to the files last
+  modified time.  For backwards compatability, the default is False and will
+  use the current date/time.
+
 v0.3.14
 -------